cleaning 'auth' route level
