+++ src/auth/endpoints/google.get.ts
@@ -1,0 +1,17 @@
+import { Request, Response } from 'express';
+import passport from 'passport';
+
+export const getAuthGoogle = passport.authenticate(
+  'google',
+  {
+    scope: ['openid', 'profile', 'email'],
+    session: false,
+  },
+  (req: Request, _res: Response) => {
+    console.log('req.user :', req.user);
+  },
+);
+
+export const getAuthGoogleCallback = passport.authenticate('google', {
+  session: false,
+});
