+++ src/auth/endpoints/index.ts
@@ -1,0 +1,5 @@
+export * from './confirmation.post';
+export * from './google.get';
+export * from './login.post';
+export * from './root.get';
+export * from './signup.post';
