+++ src/auth/google/passport-setup.ts
@@ -5,7 +5,7 @@ import { Strategy as GoogleStrategy } from 'passport-google-oauth20';
 import logger from '../../config/logger';
 import User, { IUser, IUserDocument } from '../../models/User';
 import { assertIsString } from '../../util';
-import { sendMail } from '../auth-service';
+import { sendMail } from '../services';
 import { withError } from '../../api/ApiError';
 
 interface IGoogleProfile {
